requirements.txt
streamlit
google-generativeai
pandas
numpy
scikit-learn
python-dotenv
langchain
openpyxl
